lmdb, back-mdb: use 'long' as result to speedup comparison.

+++ lmdb.h
@@ -240,7 +240,7 @@ typedef struct MDB_val {
 } MDB_val;
 
 /** @brief A callback function used to compare two keys in a database */
-typedef int  (MDB_cmp_func)(const MDB_val *a, const MDB_val *b);
+typedef long (MDB_cmp_func)(const MDB_val *a, const MDB_val *b);
 
 /** @brief A callback function used to relocate a position-dependent data item
  * in a fixed-address database.
